chore: Refactor host provider submodule
