+++ src/server/commons.py
@@ -1,13 +1,16 @@
+"""
+This file handles the configuration of the server and the global management of
+that information during runtime.
+
+That includes managing the user-defined hosts, host providers, and query cache.
+
+"""
+
 import json
 from functools import lru_cache
 from pathlib import Path
 
-from ..host_provider_router import (
-    HostProvider,
-    HostProviderRouter,
-    provider_name_map,
-)
-
+from ..host_provider.router import HostProvider, HostProviderRouter, provider_name_map
 from ..models import HostListing
 
 
@@ -48,9 +51,25 @@ def _hosts_from_json_config(hosts: list[dict]) -> list[HostListing]:
 class HostProviderRouterGlobalDep:
     """
     The global dependency that manages host provider routes.
+
+    This class is responsible for loading and storing the host provider list,
+    the host list, the routing from host requests to host providers, and the
+    query cache.
+
     """
 
     def __init__(self, json_filepath_or_dict: str | dict | Path):
+        """
+        Initialize the global dependency, loading a config.
+
+        Arguments:
+            json_filepath_or_dict (str | dict | Path): The path to the JSON
+                config file or the dictionary of the config.
+
+        Returns:
+            None
+
+        """
         if isinstance(json_filepath_or_dict, str):
             with open(json_filepath_or_dict, "r") as f:
                 config = json.load(f)
@@ -63,23 +82,67 @@ class HostProviderRouterGlobalDep:
         self.host_provider_router.validate_all_hosts(self.all_hosts)
 
     def get_uri_from_name(self, name: str) -> str | None:
+        """
+        Returns the URI of a host from its name.
+
+        Arguments:
+            name (str): The name of the host.
+
+        Returns:
+            str | None: The URI of the host, or None if it doesn't exist.
+
+        """
         for host in self.all_hosts:
             if host.name == name:
                 return host.uri
         return None
 
     def get_name_from_uri(self, uri: str) -> str | None:
+        """
+        Returns the name of a host from its URI.
+
+        Arguments:
+            uri (str): The URI of the host.
+
+        Returns:
+            str | None: The name of the host, or None if it doesn't exist.
+
+        """
         for host in self.all_hosts:
             if host.uri == uri:
                 return host.name
         return None
 
     def all_providers(self) -> list[HostProvider]:
+        """
+        Returns the list of all configured host providers.
+
+        Arguments:
+            None
+
+        Returns:
+            list[HostProvider]: The list of all configured host providers.
+
+        """
         return self.host_provider_router.all_providers()
 
 
 @lru_cache()
 def provider_router():
+    """
+    Returns the global dependency that manages host provider routes.
+
+    This function is cached, so it will only be called once during runtime.
+    (This saves the cost of the disk round-trip.)
+
+    Arguments:
+        None
+
+    Returns:
+        HostProviderRouterGlobalDep: The global dependency that manages host
+            provider routes.
+
+    """
     _here = Path(__file__).parent
     return HostProviderRouterGlobalDep(_here / "config.json")
