+++ src/server/routers/host_providers.py
@@ -1,5 +1,16 @@
+"""
+Routes for host provider management.
+
+Host providers are the delegated components that handle IO with the host graphs
+listed in the server's configuration file. The server's host provider router
+routes queries to the appropriate provider based on the string URI of the host.
+
+"""
+
 from typing import Annotated
+
 from fastapi import APIRouter, Depends
+
 from ...models import HostProviderPublicListing
 from ..commons import HostProviderRouterGlobalDep, provider_router
